+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## HEAD
+ - Forked from logstash-codec-avro_schema_registry and bulk rename to logstash-codec_avro_wapiti
+
 ## 1.1.0
  - Add support for client side TLS auth to schema registry